fix(backend): security hardening — CORS, rate limiting, debug mode
- Restrict CORS origins from wildcard (*) to configured ALLOWED_ORIGINS env var
- Add flask-limiter with 5 requests/minute rate limit on login endpoint
- Disable debug mode by default (FLASK_DEBUG env var, false by default)
- Extract duplicated token_required decorator to utils/auth_decorator.py (DRY)
- Remove dead encode_jwt function from jwt_utils.py (would break if called)
- Remove unused encode_jwt import from auth_controller.py
- Sanitize error responses (remove str(e) that leaks internals)
- Add .env.example with new security variables
- Add GIT_CONFLICT_GUIDE.md for future reference
- Add SPEC-P02-security-phase2.md spec document
