+++ backend/.env.example
@@ -1,0 +1,19 @@
+# Database
+DATABASE_URL=postgresql://user:password@localhost:5432/portfolio
+
+# Authentication
+SECRET_API_KEY=your-secret-key-here
+DEFAULT_USER=sergiopenagos
+DEFAULT_EMAIL=sergio@example.com
+DEFAULT_PASSWORD=changeme
+
+# Project Types
+DEFAULT_PROJECT_TYPE_1=Job
+DEFAULT_PROJECT_TYPE_2=Personal
+DEFAULT_PROJECT_TYPE_3=Freelance
+
+# CORS (comma-separated origins)
+ALLOWED_ORIGINS=https://gioudi.github.io,http://localhost:8080
+
+# Flask Debug (set to "true" only in local development)
+FLASK_DEBUG=false
