+++ backend/requirements.txt
@@ -1,5 +1,6 @@
 Flask==3.0.0
 Flask-CORS==4.0.0
+flask-limiter==4.1.1
 SQLAlchemy==2.0.36
 bcrypt==4.1.2
 PyJWT==2.8.0
