+++ backend/src/controllers/auth_controller.py
@@ -2,7 +2,6 @@ from flask import request, jsonify
 from services.auth_service import AuthService
 from repositories.user_repository import UserRepository
 from models.database import Session
-from utils.jwt_utils import encode_jwt
 
 
 session = Session()
@@ -17,12 +16,12 @@ def login():
         data = request.get_json()
         username = data.get('username')
         password = data.get('password')
-        
+
         token = auth_service.login(username, password)
-        
+
         if token:
             return jsonify({"message": "Login successful!", "token": token}), 200
         else:
             return jsonify({"message": "Invalid credentials!"}), 401
-    except Exception as e:
-        return jsonify({"message": "An error occurred", "error": str(e)}), 500
+    except Exception:
+        return jsonify({"message": "An error occurred"}), 500
