+++ backend/src/controllers/project_controller.py
@@ -1,31 +1,13 @@
 from flask import request, jsonify
 from services.project_service import ProjectService
-from utils.jwt_utils import decode_jwt
+from utils.auth_decorator import token_required
 
 project_service = ProjectService()
 
 
-def token_required(f):
-    def decorated(*args, **kwargs):
-        token = request.headers.get('Authorization')
-        if not token:
-            return jsonify({"message": "Token is missing!"}), 403
-        
-        # Strip "Bearer " prefix if present
-        if token.startswith("Bearer "):
-            token = token.split(" ")[1]
-        
-        user_id = decode_jwt(token)
-        if not user_id:
-            return jsonify({"message": "Token is invalid or expired!"}), 403
-        
-        return f(user_id, *args, **kwargs)
-    return decorated
-
-
 @token_required
 def create_project(user_id):
-    
+
     data = request.get_json()
     data['user_id'] = user_id
     project = project_service.create_project(data)
@@ -34,5 +16,3 @@ def create_project(user_id):
 def get_projects():
     projects = project_service.get_all_projects()
     return jsonify([project.to_dict() for project in projects])
-    
-
