+++ backend/src/controllers/project_type_controller.py
@@ -1,8 +1,8 @@
-from flask import request, jsonify
+from flask import jsonify
 from repositories.project_type_repository import ProjectTypeRepository
 from models.database import Session
 from services.project_type_service import ProjectTypeService
-from utils.jwt_utils import decode_jwt
+from utils.auth_decorator import token_required
 
 session = Session()
 
@@ -11,25 +11,7 @@ project_type_repository = ProjectTypeRepository(session)
 project_type_service = ProjectTypeService(project_type_repository)
 
 
-def token_required(f):
-    def decorated(*args, **kwargs):
-        token = request.headers.get('Authorization')
-        if not token:
-            return jsonify({"message": "Token is missing!"}), 403
-        
-        if token.startswith("Bearer "):
-            token = token.split(" ")[1]
-        
-        user_id = decode_jwt(token)
-        if not user_id:
-            return jsonify({"message": "Token is invalid or expired!"}), 403
-        
-        return f(user_id, *args, **kwargs)
-    return decorated
-
-
 @token_required
 def get_project_types(user_id):
     project_types = project_type_service.get_all_project_types()
     return jsonify([project_type.to_dict() for project_type in project_types])
-
